requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Fiona==1.8.22
GDAL==3.5.1
GeoAlchemy2==0.13.0
geopandas==0.12.2
greenlet==2.0.1
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
jupyter-client==7.4.9
jupyter-core==5.1.3
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyproj==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pyzmq==25.0.0
shapely==2.0.0
six==1.16.0
SQLAlchemy==1.4.46
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
wcwidth==0.2.6
